call run_brick directly
